JWT authentication is added

+++ src/main/java/com/cursojava/curso/controllers/UserController.java
@@ -2,11 +2,13 @@ package com.cursojava.curso.controllers;
 
 import com.cursojava.curso.dao.UserDao;
 import com.cursojava.curso.models.User;
+import com.cursojava.curso.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 // Controllers are basically used for manage our page's urls
@@ -17,13 +19,18 @@ public class UserController {
     @Autowired
     private UserDao userDao;
 
+    @Autowired
+    private JWTUtil jwtUtil;
+
     @RequestMapping(value = "api/users") // The requests have the method GET for default
-    public List<User> getUsers(){
+    public List<User> getUsers(@RequestHeader(value="Authorization") String token){
+        if(!verifyToken(token)) { return null; }
         return userDao.getUsers();
     }
 
     @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable Long id){
+    public void delete(@RequestHeader(value="Authorization") String token, @PathVariable Long id){
+        if(!verifyToken(token)) { return; }
         userDao.delete(id);
     }
 
@@ -36,6 +43,11 @@ public class UserController {
         user.setPassword(hash);
 
         userDao.register(user);
+    }
+
+    private boolean verifyToken(String token) {
+        String userId = jwtUtil.getKey(token);
+        return userId != null;
     }
 
 }
